refactor: rename file to uppercase
