+++ src/routes/router.tsx
@@ -2,7 +2,7 @@ import { createBrowserRouter } from 'react-router-dom';
 
 import Layout from '@/layouts/Layout';
 import AccountRecoveryPage from '@/pages/AccountRecoveryPage';
-import ConcatPage from 'ConcatPage';
+import ConcatPage from '@/pages/ConcatPage';
 import ErrorPage from '@/pages/ErrorPage';
 import HistoryPage from '@/pages/HistoryPage';
 import HomePage from '@/pages/HomePage';
